Fixed reflection calculation, added prism rotation with linkers.
Now the reflection is performed accurately without needing additional translation.
Prism points are updated so each reflection is performed accurately.
Size parameter for the mirror now changes a, b, c, and d values of the plane
as well. This results in accurate translation after reflection is performed.
Plane eqaution 'd' value is changed in symmetry operation.